Fixes timestamp issue on 32-bit machines and fixes sleeping bug.

+++ sensors/stereo/opencv-stereo-util.hpp
@@ -90,7 +90,7 @@ bool ParseConfigFile(string configFile, OpenCvStereoConfig *configStruct);
 
 bool LoadCalibration(string calibrationDir, OpenCvStereoCalibration *stereoCalibration);
 
-int LoadVideoFileFromDir(VideoCapture *left_video_capture, VideoCapture *right_video_capture, string video_directory, long timestamp, int video_number);
+int LoadVideoFileFromDir(VideoCapture *left_video_capture, VideoCapture *right_video_capture, string video_directory, long long timestamp, int video_number);
 
 void StopCapture(dc1394_t *dcContext, dc1394camera_t *camera);
